IK Update (Working)
Controllers now track and VR camera position is fixed as not to clip with the player body. Hands have basic setup and correspond to correct position and rotation of the user's hands now.  Body is not yet scaled properly but I am committing at this point as not to break things (again) when I mess with the body scale.